Creating user policy. Create policy base. Using Laravel Permission to check permission on policies. Checking authorization in UserController

+++ app/Http/Controllers/Api/Admin/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api\Admin;
 use App\Http\Controllers\Api\WithFilter;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\Admin\UserResource;
+use App\Models\User;
 use App\Services\UserService;
 use Illuminate\Http\Request;
 
@@ -18,6 +19,8 @@ class UserController extends Controller
      */
     public function index(): \Illuminate\Http\JsonResponse
     {
+        $this->authorize('viewAny', User::class);
+
         $filterData = $this->getValidatedFilterInputs([
             'limit' => ['nullable', 'integer', 'min:1', 'max:50'],
             'search' => ['nullable', 'string', 'max:25']
@@ -26,7 +29,7 @@ class UserController extends Controller
         return response()->json([
             'success' => true,
             'users' => UserResource::collection(
-                $this->applyFilter(\App\Models\User::class, $filterData, ['first_name'])
+                $this->applyFilter(User::class, $filterData, ['first_name'])
             )->response()->getData()
         ]);
     }
@@ -38,6 +41,8 @@ class UserController extends Controller
      */
     public function store(\App\Http\Requests\Admin\UserFormRequest $request): \Illuminate\Http\JsonResponse
     {
+        $this->authorize('create', User::class);
+
         $created = UserService::create($request->validated());
 
         return response()->json([
@@ -51,11 +56,13 @@ class UserController extends Controller
      * @param \App\Models\User $user
      * @return \Illuminate\Http\JsonResponse
      */
-    public function show(\App\Models\User $user): \Illuminate\Http\JsonResponse
+    public function show(User $user): \Illuminate\Http\JsonResponse
     {
+        $this->authorize('create', $user);
+
         return response()->json([
             'success' => true,
-            'user' => new \App\Http\Resources\Admin\UserResource($user)
+            'user' => new UserResource($user)
         ]);
     }
 
@@ -65,8 +72,10 @@ class UserController extends Controller
      * @param \App\Models\User $user
      * @return mixed|\Illuminate\Http\JsonResponse
      */
-    public function update(\App\Http\Requests\Admin\UserFormRequest $request, \App\Models\User $user)
+    public function update(\App\Http\Requests\Admin\UserFormRequest $request, User $user)
     {
+        $this->authorize('update', $user);
+
         $updated = UserService::update($user, $request->validated());
 
         return response()->json([
@@ -80,8 +89,10 @@ class UserController extends Controller
      * @param \App\Models\User $user
      * @return \Illuminate\Http\JsonResponse
      */
-    public function destroy(\App\Models\User $user)
+    public function destroy(User $user)
     {
+        $this->authorize('delete', $user);
+
         return response()->json([
             'success' => $user->delete()
         ]);
